+++ haskell/p2prc.cabal
@@ -67,11 +67,8 @@ executable p2prc
     -- Other library packages from which modules are imported.
     build-depends:      base
                       , text
-                      , aeson
-                      , process
-                      , bytestring
-                      , directory
                       , p2prc
+                      , process
 
     -- Base language which the package is written in.
     default-language: GHC2021
